Improved method mocking performance
Each time a method is mocked using $(), the code for the stub is written to disk using a unique name then included into the object and the file removed.

To improve performance, I've changed the filename generation to an MD5 hash of the code, and disabled removing the files. This allows TestBox to leverage ColdFusion template caching. This reduces the time that $() takes from ~30 ms to ~0 ms.