Fix missing raw_trace in simple.cfm
If there's a syntax error in a dependent file called from a spec, this can cause the local.thisSpec.failOrigin[ 1 ] to not have a raw_trace key, which results in a generic CF 500 error in ACF (11 & 2018 were tested). By adding a key exists check, this allows the report to render and display the exception structure allowing much easier debugging of the syntax error.